refactored audit
