+++ core/domain/audit_validators_test.py
@@ -1,0 +1,221 @@
+# coding: utf-8
+#
+# Copyright 2020 The Oppia Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS-IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Unit tests for core.domain.audit_validators."""
+
+from __future__ import absolute_import  # pylint: disable=import-only-modules
+from __future__ import unicode_literals  # pylint: disable=import-only-modules
+
+import random
+import math
+import datetime
+import time
+
+# user_models, audit_models, prod_validation_jobs_one_off, utils
+from core.domain import prod_validation_jobs_one_off
+from core.tests import test_utils
+from core.platform import models
+import feconf
+import utils
+
+(audit_models, user_models) = models.Registry.import_models([models.NAMES.audit, models.NAMES.user])
+
+USER_EMAIL = 'useremail@example.com'
+USER_NAME = 'username'
+CURRENT_DATETIME = datetime.datetime.utcnow()
+
+class RoleQueryAuditModelValidatorTests(test_utils.AuditJobsTestBase):
+
+    def setUp(self):
+        super(RoleQueryAuditModelValidatorTests, self).setUp()
+
+        self.signup(self.ADMIN_EMAIL, self.ADMIN_USERNAME)
+        self.signup(USER_EMAIL, USER_NAME)
+
+        self.admin_id = self.get_user_id_from_email(self.ADMIN_EMAIL)
+
+        admin_model = user_models.UserSettingsModel.get_by_id(self.admin_id)
+        admin_model.role = feconf.ROLE_ID_ADMIN
+        admin_model.put()
+
+        model_id = '%s.%s.%s.%s' % (
+            self.admin_id, int(math.floor(time.time())),
+            feconf.ROLE_ACTION_UPDATE, random.randint(0, 1000))
+        self.model_instance = audit_models.RoleQueryAuditModel(
+            id=model_id, user_id=self.admin_id,
+            intent=feconf.ROLE_ACTION_UPDATE, role='c', username='d')
+        self.model_instance.put()
+
+        self.job_class = (
+            prod_validation_jobs_one_off.RoleQueryAuditModelAuditOneOffJob)
+
+    def test_standard_model(self):
+        expected_output = [u'[u\'fully-validated RoleQueryAuditModel\', 1]']
+        self.run_job_and_check_output(
+            expected_output, sort=False, literal_eval=False)
+
+    def test_model_with_created_on_greater_than_last_updated(self):
+        self.model_instance.created_on = (
+            self.model_instance.last_updated + datetime.timedelta(days=1))
+        self.model_instance.put()
+        expected_output = [(
+            u'[u\'failed validation check for time field relation check '
+            'of RoleQueryAuditModel\', '
+            '[u\'Entity id %s: The created_on field has a value '
+            '%s which is greater than the value '
+            '%s of last_updated field\']]') % (
+                self.model_instance.id, self.model_instance.created_on,
+                self.model_instance.last_updated
+            )]
+        self.run_job_and_check_output(
+            expected_output, sort=False, literal_eval=False)
+
+    def test_model_with_last_updated_greater_than_current_time(self):
+        expected_output = [(
+            u'[u\'failed validation check for current time check of '
+            'RoleQueryAuditModel\', '
+            '[u\'Entity id %s: The last_updated field has a '
+            'value %s which is greater than the time when the job was run\']]'
+        ) % (self.model_instance.id, self.model_instance.last_updated)]
+
+        mocked_datetime = datetime.datetime.utcnow() - datetime.timedelta(
+            hours=13)
+        with self.mock_datetime_for_audit(mocked_datetime):
+            self.run_job_and_check_output(
+                expected_output, sort=False, literal_eval=False)
+
+    def test_model_with_non_existent_user_id(self):
+        user_models.UserSettingsModel.get(self.admin_id).delete()
+        expected_output = [(
+            u'[u\'failed validation check for user_ids field check of '
+            'RoleQueryAuditModel\', '
+            '[u"Entity id %s: based on field user_ids having value '
+            '%s, expected model UserSettingsModel with '
+            'id %s but it doesn\'t exist"]]') % (
+                self.model_instance.id, self.admin_id, self.admin_id)]
+
+        self.run_job_and_check_output(
+            expected_output, sort=False, literal_eval=False)
+
+    def test_model_with_invalid_id(self):
+        model_invalid_id = '%s.%s.%s.%s' % (
+            'a', int(math.floor(time.time())), feconf.ROLE_ACTION_UPDATE,
+            random.randint(0, 1000))
+        model_instance_with_invalid_id = audit_models.RoleQueryAuditModel(
+            id=model_invalid_id, user_id=self.admin_id,
+            intent=feconf.ROLE_ACTION_UPDATE, role='c', username='d')
+        model_instance_with_invalid_id.put()
+        expected_output = [(
+            u'[u\'fully-validated RoleQueryAuditModel\', 1]'
+        ), (
+            u'[u\'failed validation check for model id check of '
+            'RoleQueryAuditModel\', '
+            '[u\'Entity id %s: Entity id does not match regex pattern\']]'
+        ) % model_invalid_id]
+        self.run_job_and_check_output(
+            expected_output, sort=True, literal_eval=False)
+
+
+class UsernameChangeAuditModelValidatorTests(test_utils.AuditJobsTestBase):
+
+    def setUp(self):
+        super(UsernameChangeAuditModelValidatorTests, self).setUp()
+
+        self.signup(self.ADMIN_EMAIL, self.ADMIN_USERNAME)
+        self.signup(USER_EMAIL, USER_NAME)
+
+        self.admin_id = self.get_user_id_from_email(self.ADMIN_EMAIL)
+
+        admin_model = user_models.UserSettingsModel.get_by_id(self.admin_id)
+        admin_model.role = feconf.ROLE_ID_ADMIN
+        admin_model.put()
+
+        model_id = (
+            '%s.%d' % (self.admin_id, utils.get_current_time_in_millisecs()))
+        self.model_instance = audit_models.UsernameChangeAuditModel(
+            id=model_id, committer_id=self.admin_id,
+            old_username=USER_NAME, new_username='new')
+        self.model_instance.put()
+
+        self.job_class = (
+            prod_validation_jobs_one_off.UsernameChangeAuditModelAuditOneOffJob)
+
+    def test_standard_model(self):
+        expected_output = [
+            u'[u\'fully-validated UsernameChangeAuditModel\', 1]']
+        self.run_job_and_check_output(
+            expected_output, sort=False, literal_eval=False)
+
+    def test_model_with_created_on_greater_than_last_updated(self):
+        self.model_instance.created_on = (
+            self.model_instance.last_updated + datetime.timedelta(days=1))
+        self.model_instance.put()
+        expected_output = [(
+            u'[u\'failed validation check for time field relation check '
+            'of UsernameChangeAuditModel\', '
+            '[u\'Entity id %s: The created_on field has a value '
+            '%s which is greater than the value '
+            '%s of last_updated field\']]') % (
+                self.model_instance.id, self.model_instance.created_on,
+                self.model_instance.last_updated
+            )]
+        self.run_job_and_check_output(
+            expected_output, sort=False, literal_eval=False)
+
+    def test_model_with_last_updated_greater_than_current_time(self):
+        expected_output = [(
+            u'[u\'failed validation check for current time check of '
+            'UsernameChangeAuditModel\', '
+            '[u\'Entity id %s: The last_updated field has a '
+            'value %s which is greater than the time when the job was run\']]'
+        ) % (self.model_instance.id, self.model_instance.last_updated)]
+
+        mocked_datetime = datetime.datetime.utcnow() - datetime.timedelta(
+            hours=13)
+        with self.mock_datetime_for_audit(mocked_datetime):
+            self.run_job_and_check_output(
+                expected_output, sort=False, literal_eval=False)
+
+    def test_model_with_non_existent_user_id(self):
+        user_models.UserSettingsModel.get(self.admin_id).delete()
+        expected_output = [(
+            u'[u\'failed validation check for committer_ids field check of '
+            'UsernameChangeAuditModel\', '
+            '[u"Entity id %s: based on field committer_ids having value '
+            '%s, expected model UserSettingsModel with '
+            'id %s but it doesn\'t exist"]]') % (
+                self.model_instance.id, self.admin_id, self.admin_id)]
+
+        self.run_job_and_check_output(
+            expected_output, sort=False, literal_eval=False)
+
+    def test_model_with_invalid_id(self):
+        model_invalid_id = (
+            '%d.%s' % (utils.get_current_time_in_millisecs(), self.admin_id))
+        model_instance_with_invalid_id = audit_models.UsernameChangeAuditModel(
+            id=model_invalid_id, committer_id=self.admin_id,
+            old_username=USER_NAME, new_username='new')
+        model_instance_with_invalid_id.put()
+        expected_output = [(
+            u'[u\'fully-validated UsernameChangeAuditModel\', 1]'
+        ), (
+            u'[u\'failed validation check for model id check of '
+            'UsernameChangeAuditModel\', '
+            '[u\'Entity id %s: Entity id does not match regex pattern\']]'
+        ) % model_invalid_id]
+        self.run_job_and_check_output(
+            expected_output, sort=True, literal_eval=False)
+
